# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/v1/car-model-colors/": {
            "post": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Create a CarModelColor",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "CarModelColors"
                ],
                "summary": "Create a CarModelColor",
                "parameters": [
                    {
                        "description": "Create a CarModelColor",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_dto.CreateCarModelColorRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "CarModelColor response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_dto.CarModelColorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_helper.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/v1/car-model-colors/get-by-filter": {
            "post": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Get CarModelColors",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "CarModelColors"
                ],
                "summary": "Get CarModelColors",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_dto.PaginationInputWithFilter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CarModelColor response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_dto.PagedList-github_com_amirhasanpour_bama-clone-web-api_src_api_dto_CarModelColorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_helper.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/v1/car-model-colors/{id}": {
            "get": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Get a CarModelColor",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "CarModelColors"
                ],
                "summary": "Get a CarModelColor",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CarModelColor response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_dto.CarModelColorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_helper.BaseHttpResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_helper.BaseHttpResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Update a CarModelColor",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "CarModelColors"
                ],
                "summary": "Update a CarModelColor",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update a CarModelColor",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_dto.UpdateCarModelColorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CarModelColor response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_dto.CarModelColorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_helper.BaseHttpResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_helper.BaseHttpResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Delete a CarModelColor",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "CarModelColors"
                ],
                "summary": "Delete a CarModelColor",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "response",
                        "schema": {
                            "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_helper.BaseHttpResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_helper.BaseHttpResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_helper.BaseHttpResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_amirhasanpour_bama-clone-web-api_src_api_dto.CarModelColorResponse": {
            "type": "object",
            "properties": {
                "color": {
                    "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_dto.ColorResponse"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_amirhasanpour_bama-clone-web-api_src_api_dto.ColorResponse": {
            "type": "object",
            "properties": {
                "hexCode": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "github_com_amirhasanpour_bama-clone-web-api_src_api_dto.CreateCarModelColorRequest": {
            "type": "object",
            "required": [
                "carModelId",
                "colorId"
            ],
            "properties": {
                "carModelId": {
                    "type": "integer"
                },
                "colorId": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_amirhasanpour_bama-clone-web-api_src_api_dto.Filter": {
            "type": "object",
            "properties": {
                "filterType": {
                    "description": "text number",
                    "type": "string"
                },
                "from": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "type": {
                    "description": "contains notContains equals notEqual startsWith lessThan lessThanOrEqual greaterThan greaterThanOrEqual inRange endsWith",
                    "type": "string"
                }
            }
        },
...
        "github_com_amirhasanpour_bama-clone-web-api_src_api_dto.PagedList-github_com_amirhasanpour_bama-clone-web-api_src_api_dto_CarModelColorResponse": {
            "type": "object",
            "properties": {
                "hasNextPage": {
                    "type": "boolean"
                },
                "hasPreviousPage": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_dto.CarModelColorResponse"
                    }
                },
                "pageNumber": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                },
                "totalRows": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_amirhasanpour_bama-clone-web-api_src_api_dto.PaginationInputWithFilter": {
            "type": "object",
            "properties": {
                "filter": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_dto.Filter"
                    }
                },
                "pageNumber": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "sort": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_dto.Sort"
                    }
                }
            }
        },
...
        "github_com_amirhasanpour_bama-clone-web-api_src_api_dto.Sort": {
            "type": "object",
            "properties": {
                "colId": {
                    "type": "string"
                },
                "sort": {
                    "type": "string"
                }
            }
        },
        "github_com_amirhasanpour_bama-clone-web-api_src_api_dto.UpdateCarModelColorRequest": {
            "type": "object",
            "properties": {
                "carModelId": {
                    "type": "integer"
                },
                "colorId": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_amirhasanpour_bama-clone-web-api_src_api_helper.BaseHttpResponse": {
            "type": "object",
            "properties": {
                "error": {},
                "result": {},
                "resultCode": {
                    "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_helper.ResultCode"
                },
                "success": {
                    "type": "boolean"
                },
                "validationErrors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_amirhasanpour_bama-clone-web-api_src_api_validations.ValidationError"
                    }
                }
            }
        },
        "github_com_amirhasanpour_bama-clone-web-api_src_api_helper.ResultCode": {
            "type": "integer",
            "enum": [
                0,
                40001,
                40101,
                40301,
                40401,
                42901,
                42902,
                50001,
                50002
            ],
            "x-enum-varnames": [
                "Success",
                "ValidationError",
                "AuthError",
                "ForbiddenError",
                "NotFoundError",
                "LimiterError",
                "OtpLimiterError",
                "CustomRecovery",
                "InternalError"
            ]
        },
        "github_com_amirhasanpour_bama-clone-web-api_src_api_validations.ValidationError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "property": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AuthBearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}